Erotettu vihollisten liikkumislogiikka ja muu logiikka erilleen. Bonusvihollinen räjähtää, mutta räjähdysefekti on väärässä asennossa.
