+++ MuseoAR/Assets/Prefabs/bonusinvader.prefab
@@ -22,8 +22,9 @@ GameObject:
   - component: {fileID: 33976568660190446}
   - component: {fileID: 23499122748551144}
   - component: {fileID: 65930808040977560}
-  - component: {fileID: 114648949057748086}
   - component: {fileID: 54603320891848682}
+  - component: {fileID: 114748244324436848}
+  - component: {fileID: 114728386070576840}
   m_Layer: 0
   m_Name: bonusinvader
   m_TagString: enemy
@@ -112,7 +113,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.298716, y: 2.6884391, z: 1.792292}
   m_Center: {x: 0, y: 0.42215052, z: 0.04743302}
---- !u!114 &114648949057748086
+--- !u!114 &114728386070576840
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -123,10 +124,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9adf668d4ac88a84ba1892458ef04bda, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  tickSpeed: 0.8
-  movementSpeed: 4
-  movementTicks: 6
+  movementSpeed: 0
+  SinCurveMultiplier: 3
+--- !u!114 &114748244324436848
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1398113695095358}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 548b9a56c218b0246bc85652872bfece, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   deathEffect: {fileID: 1488473495704836, guid: 1b44ce44b9ab38f4ba633feac68946e9,
     type: 2}
   pointsValue: 1000
-  MoveSpeed: 0.5
