+++ MuseoAR/Assets/Scripts/Invaders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5f7d9011fe0d414f980c0f137f8c326
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
