+++ MuseoAR/Assets/Scripts/Invaders/EnemyMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 233d0e5cd7815b145ab7eaedd886bfd3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
